Add `CasePaths` trait
Enabled by default for backwards compatibility, gives folks the
opportunity to opt out of this functionality, and can be removed from
the default trait list in a future major version.
