+++ Package.resolved
@@ -1,5 +1,5 @@
 {
-  "originHash" : "f926c61158bcabafcb6a7a7a586ac4cbe0b9ac65766b4c904c1e31d058bd8f78",
+  "originHash" : "9d206b1f9cc74f0eb39838df6ef86f47f3e8dfaf16122ceeb6dc2742dcecfde2",
   "pins" : [
     {
       "identity" : "swift-case-paths",
