+++ Sources/SwiftNavigation/Traits/Sharing.swift
